docs: Ash Mudra added to list (#1191)
* adding my link to the list

thank you for this and for reviewing!

* editing directions in readme

Even with great directions, I was still a bit confused while watching the YouTube video. When I realized there were two options to accomplish the same goal, I was able to choose which I wanted. I hope adding "Option" and "1, 2" to the titles helps others. If not, I value feedback! Thanks!

* reverting changes