docs: improve readme (#1231)
* adding my link to the list

thank you for this and for reviewing!

* editing directions in readme

Even with great directions, I was still a bit confused while watching the YouTube video. When I realized there were two options to accomplish the same goal, I was able to choose which I wanted. I hope adding "Option" and "1, 2" to the titles helps others. If not, I value feedback! Thanks!

* reverting changes

* editing directions for clarity

* Updating Steps for Option 1

Adding more detailed step by step directions for completing this process in the browser. thanks!!

* fixed repository typo

Thanks for pointing that out

Co-authored-by: Tom Schmelzer <30869493+schmelto@users.noreply.github.com>